+++ app/Http/Controllers/BuildingController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\NotFoundException;
 use Illuminate\Http\Request;
 use App\Models\Building;
 
@@ -26,24 +27,45 @@ class BuildingController extends Controller
 
     public function item($id)
     {
-        $building = Building::where('id', $id)->first();
-        return $building;        
+        if (empty($id))
+        {
+            throw new NotFoundException("Id пустой!");            
+        }
+        else
+        {
+            $building = Building::where('id', $id)->with('rooms')->first();
+            return $building;        
+        }   
     } 
 
     public function update($id, Request $request)
     {
-        $building = Building::where('id', $id)->first();
-        $building->room_id = $request->room_id;
-        $building->booked = $request->booked;
-        $building->user_id = $request->user_id;
-        $building->time_start = $request->time_start;
-        $building->time_end = $request->time_end;
-        $building->save();
-        return $building;      
+        if (empty($id))
+        {
+            throw new NotFoundException("Id пустой!");            
+        }
+        else
+        {
+            $building = Building::where('id', $id)->first();
+            $building->room_id = $request->room_id;
+            $building->booked = $request->booked;
+            $building->user_id = $request->user_id;
+            $building->time_start = $request->time_start;
+            $building->time_end = $request->time_end;
+            $building->save();
+            return $building;          
+        }  
     } 
 
     public function delete($id)
     {
-        $building = Building::where('id', $id)->delete();        
+        if (empty($id))
+        {
+            throw new NotFoundException("Id пустой!");            
+        }
+        else
+        {
+            $building = Building::where('id', $id)->delete();          
+        }     
     } 
 }
